backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
psycopg[binary]==3.1.19
django-environ==0.11.2
celery==5.4.0
redis==5.0.4
gunicorn==22.0.0
whitenoise==6.6.0
django-cors-headers==4.3.1
